# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "description": "Get list of orders for a specific restaurant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List restaurant orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant ID",
                        "name": "restaurant_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Order Status Filter",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "description": "Create a new order with items and options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models_restaurant.CreateOrder"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/orders/{id}": {
            "get": {
                "description": "Get detailed information about a specific order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/orders/{id}/cancel": {
            "post": {
                "description": "Cancel an existing order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/orders/{id}/status": {
            "put": {
                "description": "Update the status of a specific order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/payments/dine/{order_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new Payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new Payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "models_restaurant.CreateOrder": {
            "type": "object",
            "required": [
                "customer_name",
                "customer_phone",
                "items",
                "order_type",
                "payment_type",
                "restaurant_id"
            ],
            "properties": {
                "customer_email": {
                    "type": "string"
                },
                "customer_name": {
                    "type": "string"
                },
                "customer_phone": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models_restaurant.CreateOrderItem"
                    }
                },
                "notes": {
                    "type": "string"
                },
                "order_type": {
                    "type": "string",
                    "enum": [
                        "DINEIN",
                        "PICKUP",
                        "DELIVERY"
                    ]
                },
                "payment_type": {
                    "type": "string",
                    "enum": [
                        "online",
                        "onsite"
                    ]
                },
                "restaurant_id": {
                    "type": "string"
                }
            }
        },
        "models_restaurant.CreateOrderItem": {
            "type": "object",
            "required": [
                "menu_item_id",
                "quantity"
            ],
            "properties": {
                "item_option_id": {
                    "type": "string"
                },
                "item_option_name": {
                    "type": "string"
                },
                "menu_item_id": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        }
    }
}